+++ TrelloDotNet/TrelloDotNet/Model/Member.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Text.Json.Serialization;
+using TrelloDotNet.Control;
 
 namespace TrelloDotNet.Model
 {
@@ -38,11 +39,31 @@ namespace TrelloDotNet.Model
         public string Initials { get; private set; }
 
         /// <summary>
-        /// The url of this member's avatar
+        /// The base url of this member's avatar
         /// </summary>
         [JsonPropertyName("avatarUrl")]
         [JsonInclude]
         public string AvatarUrl { get; private set; }
+
+        /// <summary>
+        /// 30x30 Pixel version of the Members Avatar Image
+        /// </summary>
+        public string AvatarUrl30 => AvatarUrl != null ? $"{AvatarUrl}/30.png" : null;
+
+        /// <summary>
+        /// 50x50 Pixel version of the Members Avatar Image
+        /// </summary>
+        public string AvatarUrl50 => AvatarUrl != null ? $"{AvatarUrl}/50.png" : null;
+
+        /// <summary>
+        /// 170x170 Pixel version of the Members Avatar Image
+        /// </summary>
+        public string AvatarUrl170 => AvatarUrl != null ? $"{AvatarUrl}/170.png" : null;
+
+        /// <summary>
+        /// Original version of the Members Avatar Image
+        /// </summary>
+        public string AvatarUrlOriginal => AvatarUrl != null ? $"{AvatarUrl}/original.png" : null;
 
         /// <summary>
         /// Whether the user has confirmed their email address for their account.
@@ -50,6 +71,21 @@ namespace TrelloDotNet.Model
         [JsonPropertyName("confirmed")]
         [JsonInclude]
         public bool Confirmed { get; private set; }
+
+        /// <summary>
+        /// The Email of the member (only populated if field is included in GetMemberOptions)
+        /// </summary>
+        [JsonPropertyName("email")]
+        [JsonInclude]
+        public string Email { get; private set; }
+
+        /// <summary>
+        /// The Type of the Members (admin, normal, observer) [only populated if field is included in GetMemberOptions]
+        /// </summary>
+        [JsonPropertyName("memberType")]
+        [JsonInclude]
+        [JsonConverter(typeof(EnumViaJsonPropertyConverter<MembershipType>))]
+        public MembershipType MemberType { get; private set; }
 
         internal static Member CreateDummy()
         {